added projects & goals screens, improved home charts, fixed login navigation & other fixed & additions

+++ app/src/main/java/com/stefdp/hackatime/network/hackatimeapi/models/Language.kt
@@ -10,8 +10,8 @@ data class Language(
     override val name: String,
     @SerializedName("total_seconds") override val totalSeconds: Double,
     override val text: String,
-    override val hours: Int,
-    override val minutes: Int,
+    override val hours: Double,
+    override val minutes: Double,
     override val percent: Double,
     override val digital: String,
 ) : GeneralStat, Parcelable
